+++ lib/settings/pages/general_settings_page.dart
@@ -1,20 +1,25 @@
+import 'dart:io';
+
 import 'package:dynamic_color/dynamic_color.dart';
 import 'package:flutter/material.dart';
 
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 import 'package:lemmy_api_client/v3.dart';
+import 'package:permission_handler/permission_handler.dart';
 
 import 'package:thunder/core/enums/local_settings.dart';
 import 'package:thunder/core/enums/nested_comment_indicator.dart';
 import 'package:thunder/core/singletons/preferences.dart';
 import 'package:thunder/settings/widgets/list_option.dart';
+import 'package:thunder/settings/widgets/settings_list_tile.dart';
 import 'package:thunder/settings/widgets/toggle_option.dart';
 import 'package:thunder/shared/comment_sort_picker.dart';
 import 'package:thunder/shared/sort_picker.dart';
 import 'package:thunder/thunder/bloc/thunder_bloc.dart';
 import 'package:thunder/utils/bottom_sheet_list_picker.dart';
 import 'package:thunder/utils/constants.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 class GeneralSettingsPage extends StatefulWidget {
   const GeneralSettingsPage({super.key});
@@ -39,6 +44,7 @@ class _GeneralSettingsPageState extends State<GeneralSettingsPage> with SingleTi
   // General Settings
   bool scrapeMissingPreviews = false;
   bool openInExternalBrowser = false;
+  bool openInReaderMode = false;
   bool useDisplayNames = true;
   bool markPostReadOnMediaView = false;
   bool showInAppUpdateNotification = false;
@@ -114,6 +120,10 @@ class _GeneralSettingsPageState extends State<GeneralSettingsPage> with SingleTi
       case LocalSettings.openLinksInExternalBrowser:
         await prefs.setBool(LocalSettings.openLinksInExternalBrowser.name, value);
         setState(() => openInExternalBrowser = value);
+        break;
+      case LocalSettings.openLinksInReaderMode:
+        await prefs.setBool(LocalSettings.openLinksInReaderMode.name, value);
+        setState(() => openInReaderMode = value);
         break;
       case LocalSettings.useDisplayNamesForUsers:
         await prefs.setBool(LocalSettings.useDisplayNamesForUsers.name, value);
@@ -280,6 +290,7 @@ class _GeneralSettingsPageState extends State<GeneralSettingsPage> with SingleTi
 
       // Links
       openInExternalBrowser = prefs.getBool(LocalSettings.openLinksInExternalBrowser.name) ?? false;
+      openInReaderMode = prefs.getBool(LocalSettings.openLinksInReaderMode.name) ?? false;
       scrapeMissingPreviews = prefs.getBool(LocalSettings.scrapeMissingPreviews.name) ?? false;
 
       // Notification Settings
@@ -318,9 +329,9 @@ class _GeneralSettingsPageState extends State<GeneralSettingsPage> with SingleTi
   @override
   Widget build(BuildContext context) {
     final ThemeData theme = Theme.of(context);
-
+    final l10n = AppLocalizations.of(context)!;
     return Scaffold(
-      appBar: AppBar(title: const Text('General'), centerTitle: false),
+      appBar: AppBar(title: Text(l10n.general), centerTitle: false),
       body: isLoading
           ? const Center(child: CircularProgressIndicator())
           : SingleChildScrollView(
@@ -335,7 +346,7 @@ class _GeneralSettingsPageState extends State<GeneralSettingsPage> with SingleTi
                         Padding(
                           padding: const EdgeInsets.only(left: 4, bottom: 8.0),
                           child: Text(
-                            'Feed',
+                            l10n.feed,
                             style: theme.textTheme.titleLarge,
                           ),
                         ),
@@ -439,14 +450,14 @@ class _GeneralSettingsPageState extends State<GeneralSettingsPage> with SingleTi
                         Padding(
                           padding: const EdgeInsets.only(left: 4.0, bottom: 8.0),
                           child: Text(
-                            'Posts',
+                            l10n.posts,
                             style: theme.textTheme.titleLarge,
                           ),
                         ),
                         Padding(
                           padding: const EdgeInsets.all(6.0),
                           child: Text(
-                            'These settings apply to the cards in the main feed, actions are always available when actually opening posts.',
+                            l10n.settingsFeedCards,
                             style: TextStyle(
                               color: theme.colorScheme.onBackground.withOpacity(0.75),
                             ),
@@ -457,7 +468,7 @@ class _GeneralSettingsPageState extends State<GeneralSettingsPage> with SingleTi
                         ),
                         ToggleOption(
                           description: LocalSettings.useCompactView.label,
-                          subtitle: 'Enable for small posts, disable for big.',
+                          subtitle: l10n.useCompactView,
                           value: useCompactView,
                           iconEnabled: Icons.crop_16_9_rounded,
                           iconDisabled: Icons.crop_din_rounded,
@@ -570,7 +581,7 @@ class _GeneralSettingsPageState extends State<GeneralSettingsPage> with SingleTi
                         ),
                         ToggleOption(
                           description: LocalSettings.dimReadPosts.label,
-                          subtitle: 'Read posts will be grayed out',
+                          subtitle: l10n.dimReadPosts,
                           value: dimReadPosts,
                           iconEnabled: Icons.chrome_reader_mode,
                           iconDisabled: Icons.chrome_reader_mode_outlined,
@@ -595,7 +606,7 @@ class _GeneralSettingsPageState extends State<GeneralSettingsPage> with SingleTi
                         Padding(
                           padding: const EdgeInsets.only(left: 4.0, bottom: 8.0),
                           child: Text(
-                            'Comments',
+                            l10n.comments,
                             style: theme.textTheme.titleLarge,
                           ),
                         ),
@@ -620,7 +631,7 @@ class _GeneralSettingsPageState extends State<GeneralSettingsPage> with SingleTi
                           icon: Icons.comment_bank_rounded,
                           onChanged: (_) {},
                           customListPicker: CommentSortPicker(
-                            title: 'Comment Sort Type',
+                            title: l10n.commentSortType,
                             onSelect: (value) {
                               setPreferences(LocalSettings.defaultCommentSortType, value.payload.name);
                             },
@@ -680,7 +691,7 @@ class _GeneralSettingsPageState extends State<GeneralSettingsPage> with SingleTi
                                   key: ValueKey(enableCommentNavigation),
                                   child: ToggleOption(
                                     description: LocalSettings.combineNavAndFab.label,
-                                    subtitle: 'Floating Action Button will be shown between navigation buttons',
+                                    subtitle: l10n.combineNavAndFab,
                                     value: combineNavAndFab,
                                     iconEnabled: Icons.join_full_rounded,
                                     iconDisabled: Icons.join_inner_rounded,
@@ -701,13 +712,13 @@ class _GeneralSettingsPageState extends State<GeneralSettingsPage> with SingleTi
                         Padding(
                           padding: const EdgeInsets.only(left: 4.0, bottom: 8.0),
                           child: Text(
-                            'Links',
+                            l10n.link(3),
                             style: theme.textTheme.titleLarge,
                           ),
                         ),
                         ToggleOption(
                           description: LocalSettings.scrapeMissingPreviews.label,
-                          subtitle: 'Enabling will have a performance hit',
+                          subtitle: l10n.scrapeMissingPreviews,
                           value: scrapeMissingPreviews,
                           iconEnabled: Icons.image_search_rounded,
                           iconDisabled: Icons.link_off_rounded,
@@ -720,6 +731,27 @@ class _GeneralSettingsPageState extends State<GeneralSettingsPage> with SingleTi
                           iconDisabled: Icons.link_rounded,
                           onToggle: (bool value) => setPreferences(LocalSettings.openLinksInExternalBrowser, value),
                         ),
+
+                        if (Platform.isIOS)
+                          ToggleOption(
+                            description: LocalSettings.openLinksInReaderMode.label,
+                            value: openInReaderMode,
+                            iconEnabled: Icons.menu_book_rounded,
+                            iconDisabled: Icons.menu_book_rounded,
+                            onToggle: (bool value) => setPreferences(LocalSettings.openLinksInReaderMode, value),
+                          ),
+                        // TOOD:(open_lemmy_links_walkthrough) maybe have the open lemmy links walkthrough here
+                        if (Platform.isAndroid)
+                          SettingsListTile(
+                            icon: Icons.add_link,
+                            widget: const Icon(
+                              Icons.arrow_forward_ios,
+                              size: 16,
+                            ),
+                            onTap: () => openAppSettings(),
+                            subtitle: l10n.allowOpenSupportedLinks,
+                            description: l10n.openByDefault,
+                          ),
                       ],
                     ),
                   ),
@@ -732,12 +764,12 @@ class _GeneralSettingsPageState extends State<GeneralSettingsPage> with SingleTi
                         Padding(
                           padding: const EdgeInsets.only(left: 4.0, bottom: 8.0),
                           child: Text(
-                            'User Profiles',
+                            l10n.userProfiles,
                             style: theme.textTheme.titleLarge,
                           ),
                         ),
                         ToggleOption(
-                          description: 'Display User Scores (Karma)',
+                          description: l10n.displayUserScore,
                           value: scoreCounters,
                           iconEnabled: Icons.score_rounded,
                           iconDisabled: Icons.score_rounded,
@@ -755,7 +787,7 @@ class _GeneralSettingsPageState extends State<GeneralSettingsPage> with SingleTi
                         Padding(
                           padding: const EdgeInsets.only(left: 4.0, bottom: 8.0),
                           child: Text(
-                            'Notifications',
+                            l10n.notifications(3),
                             style: theme.textTheme.titleLarge,
                           ),
                         ),
@@ -778,7 +810,7 @@ class _GeneralSettingsPageState extends State<GeneralSettingsPage> with SingleTi
                         Padding(
                           padding: const EdgeInsets.only(left: 4.0, bottom: 8.0),
                           child: Text(
-                            'Import/Export Settings',
+                            l10n.importExportSettings,
                             style: theme.textTheme.titleLarge,
                           ),
                         ),
@@ -793,7 +825,7 @@ class _GeneralSettingsPageState extends State<GeneralSettingsPage> with SingleTi
                               const Icon(Icons.settings_rounded),
                               const SizedBox(width: 8.0),
                               Text(
-                                'Save Settings',
+                                l10n.saveSettings,
                                 style: TextStyle(color: theme.colorScheme.onPrimaryContainer),
                               ),
                             ],
@@ -814,7 +846,7 @@ class _GeneralSettingsPageState extends State<GeneralSettingsPage> with SingleTi
                               const Icon(Icons.import_export_rounded),
                               const SizedBox(width: 8.0),
                               Text(
-                                'Import Settings',
+                                l10n.importSettings,
                                 style: TextStyle(color: theme.colorScheme.onPrimaryContainer),
                               ),
                             ],