+++ CHANGELOG.md
@@ -5,17 +5,22 @@
 - Instance autocomplete when logging in - contribution from @micahmo
 - Show trending communities on search page - contribution from @micahmo
 - Added new post sharing options - contribution from @micahmo
-- Added Read All in inbox. - contribution from @ggriffo
-- Open links to posts in the app instead of browser - contribution from @micahmo
+- Added Read All in inbox. - contribution from @ggichure
+- Open links to posts and comments in the app instead of browser - contribution from @micahmo
 - Added support for lemmy 0.19.x authentication - contribution from @micahmo
 - Added support for accessibility profiles in settings - contribution from @micahmo
 - Added option to enable/disable full screen navigation swipe gesture to go back (applies when LTR gestures are disabled)
+- Introduced support for reporting comments. - contribution from @ggichure
+- Added ability to create post from home feed - contribution from @micahmo
+- Added option to enter reader mode when tapping on a link in iOS
+- Added support to open Lemmy links in app. Android only.  - contribution from @ggichure
 
 ### Changed
 - Collapsed comments are easier to expand - contribution from @micahmo
 - Show up to 99 before adding + in the unread count - contribution from @micahmo
 - Migrate from old BottomNavigationBar to NavigationBar - contribution from @ggichure
 - Adjusted logic to allow for instant switching of sort types, refreshing feed, and switching communities/pages in drawer
+- Removed overscroll effect on Android when reduce animation setting is enabled
 
 ### Fixed
 - Handle issue where failing to retrieve image dimensions blocks post loading - contribution from @Fmstrat